requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
asgiref==3.10.0
sqlparse==0.5.3
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.9
dj-database-url==2.3.0
